+++ config/vitest.ts
@@ -5,8 +5,8 @@ import { env } from 'node:process';
 import { defineConfig } from 'vitest/config';
 
 // Make sure the output of the CLI is in color, so that it matches the
-// snapshots.
-env.FORCE_COLOR = 2;
+// snapshots.  This needs to be set as a string to work correctly.
+env.FORCE_COLOR = '2';
 
 export default defineConfig({
   test: {