+++ source/types.ts
@@ -85,6 +85,7 @@ export declare interface Options {
   '--ssl-cert': Path;
   '--ssl-key': Path;
   '--ssl-pass': string;
+  '--token': string;
 }
 
 // The arguments passed to the CLI (the options + the positional arguments)